Only wait 1 second max on logout/shutdown
This fixes an issue which was reproducible with dropbox in Linux Mint 17.x.
dropbox was registered as an XSMP app.

In the exit phase, cinnamon-session would tell it xsmp_stop() and that would issue an SmsDie() call on it.

That would hang for a good 10 seconds or so and we would then receive the following error:

cinnamon-session[1261]: DEBUG(+): CsmXsmpServer: ice_io_error_handler (0x242bc10)
cinnamon-session[1261]: DEBUG(+): CsmXSMPClient: IceProcessMessagesIOError on '0x23d3910 [/home/username/.dropbox-dist/dropbox-lnx.x86_64-3.10.7/dropbox 108c3699a2c71ea81144475366310387500000012610062]'

There was also already a timeout in place for the phase to end, but it was the same timeout for all phases and it was a ridiculously large 30 seconds.

By the time we get to the EXIT phase, all apps were already given the chance to inhibit the logout/shutdown, and the only thing that is at play here is for them to actually exit.

It's not ok for apps to delay this particular phase, so we're applying a minimal timeout value of just 1 second.