Restart dbus user-session if last graphical session (#99)
* system: add api for detecting if this is the last session for a user

https://bugzilla.gnome.org/show_bug.cgi?id=764029

* manager: restart user bus if last graphical session

There are desktop services (such as goa-daemon, e-d-s, etc)
that don't open the display, but rely on dbus-daemon to scope
the session. These days dbus-daemon is a user bus, not a sesssion,
bus which leaves these services alive after log out.

This commit checks to see if we're the last desktop session for the
user at log out time, and if so, restarts the dbus daemon. This
will lead to existing clients getting booted and die, but allow user
bus clients that want to outlive a session to stick around if they so
desire.

Longer term, clients should stop relying on the session bus to define
their lifetime.

https://bugzilla.gnome.org/show_bug.cgi?id=764029