Port to gdbus (#125)
* all: Port to gdbus.

The main purpose of this is to make things a bit easier to maintain and develop
on, without needing to use the obsolete dbus-glib api.

- Removes dependencies on dbus-glib
- Simplifies code a bit
- Adds a test to simulate a session proxy for monitoring properties from a
  client perspective.

* csm-manager.c: Don't let inhibitors stop a forced logout.

When confronted with an inhibit dialog and 'logout anyway' was selected,
the session was still waiting till the 30 second timeout before actually
terminating.  This lets the session end immediately if we want to force
it.

* csm-autostart-app.c: Use GInitable to avoid unref during construction
warning when an autostart file fails to be read.

* csm-manager.c: Push property changes immediately when adding or removing
inhibitors.

Without this, the InhibitorAdded and InhibitorRemoved signals are received
by clients before the InhibitedActions cached value is updated.

Users like the cinnamon inhibitor applet that rely on these signals would end
up getting a stale InhibitedActions value.

* csm-manager.c: Disable dbus restart at the end of the session.

This was broken previously, but when it actually works, it resolves
some issues with dbus-involved processes not terminating when the
session ends.  However, with it working, it now causes certain other
programs to not end properly.

Keep it non-working for now.