csm-presence.c: Don't start cinnamon-screensaver during construction.
We only listen to the screensaver for ActiveChanged here.  Also, if
it's started this early, it interferes with any custom screensaver
that has been defined.

There are multiple  other triggers further on that will start the
screensaver (including its desktop file), if there is no custom one
configured.

ref:
linuxmint/cinnamon-screensaver@095b81385f
https://github.com/linuxmint/cinnamon-screensaver/pull/323