csm-logout-dialog.c: Grab focus after the widget is shown.
Cinnamon-session is a gtk program, but its only window is the
logout/shutdown dialog. The first time the dialog is shown, it
gets focus by default. If the dialog is cancelled and a new
one made later, the last timestamp from the previous dialog is
used to determine whether or not to give it focus. This old
timestamp will usually fail (since there are almost always
intervening user events with other programs between dialogs),
and the dialog will be denied focus.

This can be reproduced in older cinnamon versions.

The problem was originally worked around by making an explicit grab
after the dialog was shown, but this was removed in bbc4fdbdd074.
This commit re-adds a callback, but uses gdk_window_focus, and
doesn't interfere with the screen locker (which was the reason
for the commit that removed this callback originally).