wayland: Add DisplayServer phase
Cinnamon is still (unfortunately) a monolithic WM/panel/DE process.

In Xorg we want it to be started after the settings daemon, so that
xsettings/xrandr etc.. get to do their magic before we start dealing
with displays and Gtk widgets.

In Wayland, we can't have Gtk initialized before the compositor is up.
So in this case we need to start Cinnamon before the settings daemon.

The settings daemon are started during the Initialization phase.

In Xorg we'll continue to start Cinnamon in the WindowManager phase (
i.e. after).

In Wayland we're introducing a new phase to start Cinnamon, called the
DisplayServer phase, which happens before.