wayland: Reimplement logout/inhibit dialogs as a separate process
We can't use Gtk in the session manager process, so make a
separate dialog to interact with it.

After the user chooses an action, the dialog stays long enough
to list any inhibitors, and displays them if needed. Once
the dialog is in this state, you can only continue/cancel (like
the existing inhibit dialog).

Only the session manager will launch the logout *dialog*. If
someone calls cinnamon-session-quit from somewhere, it will run
briefly and forward the command to the public session manager
interface. Then the manager will spawn the real dialog.